NEWS-AGGREGATOR/requirements.txt
annotated-types==0.7.0
APScheduler==3.11.0
beautifulsoup4==4.13.4
bidict==0.23.1
blinker==1.9.0
blis==1.3.0
breadability==0.1.20
catalogue==2.0.10
certifi==2025.8.3
chardet==5.2.0
charset-normalizer==3.4.3
click==8.2.1
cloudpathlib==0.21.1
confection==0.1.5
cymem==2.0.11
dnspython==2.7.0
docopt==0.6.2
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
eventlet==0.40.2
feedparser==6.0.11
filelock==3.18.0
Flask==3.1.1
Flask-APScheduler==1.13.1
Flask-SocketIO==5.5.1
fsspec==2025.7.0
greenlet==3.2.4
h11==0.16.0
hf-xet==1.1.7
huggingface-hub==0.34.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
langcodes==3.5.0
language_data==1.3.0
lxml==6.0.0
marisa-trie==1.2.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.13
networkx==3.5
nltk==3.9.1
numpy==2.3.2
packaging==25.0
pillow==11.3.0
preshed==3.0.10
pycountry==24.6.1
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-engineio==4.12.2
python-socketio==5.13.0
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.4
rich==14.1.0
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.16.1
sentence-transformers==5.1.0
setuptools==80.9.0
sgmllib3k==1.0.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
smart_open==7.3.0.post1
soupsieve==2.7
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
sumy==0.11.0
sympy==1.14.0
thinc==8.3.6
threadpoolctl==3.6.0
tokenizers==0.21.4
torch==2.8.0
tqdm==4.67.1
transformers==4.55.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzlocal==5.3.1
urllib3==2.5.0
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.1.3
wrapt==1.17.3
wsproto==1.2.0
